Moved SpatialIndexSerde to serde dir->SpatialIndex Removed unused GeometrySerde object from SpatialIndexSerde (eliminate the requirement to handle WKB and Shape GeometrySerde)
